% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy}
\usage{
accuracy(data = NULL, obs, pred, tidy = FALSE, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (character | factor).}

\item{pred}{Vector with predicted values (character | factor).}

\item{tidy}{Logical operator (TRUE/FALSE) to decide the type of return. TRUE returns a data.frame, FALSE returns a list (default).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
an object of class \code{numeric}.
}
\description{
It estimates the accuracy for a nominal/categorical predicted-observed dataset.
}
\details{
Accuracy refers to a measure of the degree to which the predictions of a
model matches the reality being modeled. The classification accuracy is calculated
as the ratio between the number of correctly classified objects with respect to
the total number of objects. It is bounded between 0 and 1. The closer to 1 the better.
Values towards zero indicate low accuracy of predictions. It can be also expressed
as percentage if multiplied by 100. It is estimated at a global level (not
at the class level).
Accuracy is directly related to the error_rate, since accuracy = 1 – error_rate' .
For the formula and more details, see
\href{https://adriancorrendo.github.io/metrica/articles/available_metrics.html}{online-documentation}
}
\examples{
\donttest{
set.seed(123)
# Two-class
binomial_case <- data.frame(labels = sample(c("True","False"), 100, 
replace = TRUE), predictions = sample(c("True","False"), 100, replace = TRUE))
# Multi-class
multinomial_case <- data.frame(labels = sample(c("Red","Blue", "Green"), 100, 
replace = TRUE), predictions = sample(c("Red","Blue", "Green"), 100, replace = TRUE) )

# Get accuracy estimate for two-class case
accuracy(data = binomial_case, obs = labels, pred = predictions, tidy = TRUE)

# Get accuracy estimate for multi-class case
accuracy(data = multinomial_case, obs = labels, pred = predictions, tidy = TRUE)
}
}
\references{
(2017) Accuracy. \emph{In: Sammut C., Webb G.I. (eds) Encyclopedia of Machine Learning and Data Mining}
\emph{Springer, Boston, MA.} \doi{10.1007/978-1-4899-7687-1_3}
}
