% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B1.sma.R
\name{B1.sma}
\alias{B1.sma}
\title{B1.sma}
\usage{
B1.sma(obs, pred, orientation = "PO", na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{orientation}{Argument of class string specifying the axis
orientation, PO for predicted vs observed, and OP for
observed vs predicted. Default is orientation = "PO".}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na_rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Slope of standardized major axis regression
(symmetric).
}
\details{
Calculates the slope (B1) for the bivariate
Predicted-Observed dataset following the standardized major axis
regression, Warton et al. (2006).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 9)
B1.sma(obs = X, pred = Y)
}
}
