% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barley}
\alias{barley}
\title{Barley grain number}
\format{
This data frame has 69 rows and the following 2 columns:
\describe{
\item{pre}{predicted values}
\item{obs}{observated values}
}
}
\source{
\url{https://github.com/adriancorrendo/metrica}
}
\usage{
barley
}
\description{
This example dataset is a set of APSIM simulations of
barley grain number (x1000 grains per squared meter), which presents high
accuracy but medium precision. The experimental trials come from 11
site-years in 2 countries (Australia, and New Zealand).
These data correspond to the latest, up-to-date, documentation and
validation of version number 2020.03.27.4956. Data available at:
\url{https://doi.org/10.7910/DVN/EJS4M0}.
Further details can be found at the official APSIM Next Generation
website:
\url{https://APSIMnextgeneration.netlify.app/modeldocumentation/}
}
\keyword{datasets}
