% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland.altman.plot.R
\name{bland.altman.plot}
\alias{bland.altman.plot}
\title{bland.altman.plot}
\usage{
bland.altman.plot(obs, pred)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}
}
\value{
Object of class \code{ggplot}.
}
\description{
Bland-Altman plot.
}
\details{
Creates a scatter plot of the difference between
observed and predicted values (obs-pred) vs. observed values.
}
\examples{
\dontrun{
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 10)
blandaltman.plot(obs = X, pre = Y)
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{aes}}
}
