% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCC.R
\name{CCC}
\alias{CCC}
\title{CCC}
\usage{
CCC(obs, pre)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pre}{Vector with predicted values (numeric).}
}
\value{
Element of class \code{double}.
}
\description{
Concordance Correlation Coefficient.
}
\details{
Calculates the CCC for Predicted-Observed dataset.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 9)
CCC(obs = X, pre = Y)
}
}
