% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chickpea}
\alias{chickpea}
\title{Chickpea dry mass}
\format{
This data frame has 39 rows and the following 2 columns:
\describe{
\item{pre}{predicted values}
\item{obs}{observated values}
}
}
\source{
\url{https://github.com/adriancorrendo/metrica}
}
\usage{
chickpea
}
\description{
This example dataset is a set of APSIM simulations of
chickpea aboveground dry mass (kg per hectare), which exhibits low
accuracy and medium-low precision as it represents a model still under
development. The experimental trials come from 7 site-years in
3 countries (Australia, India, and New Zealand).
These data correspond to the latest, up-to-date, documentation and
validation of version number 2020.03.27.4956. Data available at:
\url{https://doi.org/10.7910/DVN/EJS4M0}.
Further details can be found at the official APSIM Next Generation
website:
\url{https://apsimnextgeneration.netlify.app/modeldocumentation/}
}
\keyword{datasets}
