% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E1.R
\name{E1}
\alias{E1}
\title{E1}
\usage{
E1(data = NULL, obs, pred, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frama contaning the data.}

\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Legates and McCabe's Model Efficiency (E1).
}
\details{
Calculates the E1 for a Predicted-Observed dataset
following Legates and McCabee (1990).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 9)
E1(obs = X, pred = Y)
}
}
