% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erel.R
\name{Erel}
\alias{Erel}
\title{Erel}
\usage{
Erel(obs, pred, na_rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}
}
\value{
Element of class \code{numeric}.
}
\description{
Krause Relative Model Efficiency (Erel).
}
\details{
Calculates the Erel for a Predicted-Observed dataset
following Krause et al. (2005).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 9)
Erel(obs = X, pred = Y)
}
}
