% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscore.R
\name{fscore}
\alias{fscore}
\title{F-score}
\usage{
fscore(data = NULL, obs, pred, B = 1, tidy = FALSE, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (character | factor).}

\item{pred}{Vector with predicted values (character | factor).}

\item{B}{Numeric value indicating the weight (a.k.a. β or beta) to be applied to the
relationship between recall and precision. β < 1 weights more precision than recall.
β > 1 gives B times more importance to recall than precision. Default: 1.}

\item{tidy}{Logical operator (TRUE/FALSE) to decide the type of return. TRUE
returns a data.frame, FALSE returns a list; Default : FALSE.}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
an object of class \code{numeric}.
}
\description{
It estimates the F-score for a nominal/categorical predicted-observed dataset.
}
\details{
The F-score (or F-measure) is the harmonic mean of precision and recall.
The universal version of F-score employs a coefficient β, by which can be the
precision-recall ratio tuned.

For binomial cases, F-score  =  TP / (TP + 0.5*(FP + FN))

It is bounded between 0 and 1.
The closer to 1 the better. Values towards zero indicate low performance.
For the formula and more details, see
\href{https://adriancorrendo.github.io/metrica/articles/available_metrics.html}{online-documentation}
}
\examples{
\donttest{
set.seed(123)
# Two-class
binomial_case <- data.frame(labels = sample(c("True","False"), 100, replace = TRUE), 
predictions = sample(c("True","False"), 100, replace = TRUE))
# Multi-class
multinomial_case <- data.frame(labels = sample(c("Red","Blue", "Green"), 100, replace = TRUE),
predictions = sample(c("Red","Blue", "Green"), 100, replace = TRUE)    )

# Get F-score estimate for two-class case
fscore(data = binomial_case, obs = labels, pred = predictions, tidy = TRUE)

# Get F-score estimate for each class for the multi-class case
fscore(data = multinomial_case, obs = labels, pred = predictions, tidy = TRUE)

# Get F-score estimate for the multi-class case at a global level
fscore(data = multinomial_case, obs = labels, pred = predictions, tidy = TRUE)
}
}
\references{
Goutte, C., Gaussier, E. (2005).
A Probabilistic Interpretation of Precision, Recall and F-Score, with Implication for Evaluation.
\emph{In: D.E. Losada and J.M. Fernandez-Luna (Eds.): ECIR 2005}
\emph{. Advances in Information Retrieval LNCS 3408, pp. 345–359, 2.}
\emph{Springer-Verlag Berlin Heidelberg.} \doi{10.1007/978-3-540-31865-1_25}
}
