% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.apsim.out.R
\name{import.apsim.out}
\alias{import.apsim.out}
\title{import.apsim.out}
\usage{
import.apsim.out(filepath)
}
\arguments{
\item{filepath}{Argument to specify the location path to the *.out file.}
}
\value{
Element of class \code{data.frame}.
}
\description{
Function to read/import *.out files and display them as data frames.
}
\details{
This function will generate a data frame from an APSIM Classic *.out file.
}
\examples{
\dontrun{
#' path <- ##
apsim.out <- import.apsim.out(filepath = path)
}
}
