% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.apsim.R
\name{import.apsim}
\alias{import.apsim}
\title{import.apsim}
\usage{
import.apsim(filepath)
}
\arguments{
\item{filepath}{Argument to specify the location path to the *.out file.}
}
\value{
Element of class \code{data.frame}.
}
\description{
Function to read/import *.out files and display them as data frames.
}
\details{
This function will generate a data frame from an APSIM Classic *.out file.
}
\examples{
\dontrun{
# Example out from apsimx package repo (by F. Miguez)
path <- ##
apsim.out <- import.apsim(filepath = path)
}
}
