% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_apsim_out.R
\name{import_apsim_out}
\alias{import_apsim_out}
\title{import_apsim_out}
\usage{
import_apsim_out(filepath)
}
\arguments{
\item{filepath}{Argument to specify the location path to the *.out file.}
}
\value{
Element of class \code{data.frame}.
}
\description{
Function to import *.out files generated by APSIM Classic.
}
\details{
This function will generate a data frame object from an APSIM Classic *.out file.
}
\note{
Note: It imports one file at a time.
}
\examples{
\dontrun{
# Pull the git repo and use the following filepath
path <- "tests/testthat/examples/soybean.out"
soybean.out <- import_apsim_out(filepath = path)
head(soybean.out)
}
}
\seealso{
\code{\link[apsimx]{read_apsim}}
}
