% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqRMSE.R
\name{iqRMSE}
\alias{iqRMSE}
\title{iqRMSE}
\usage{
iqRMSE(obs, pred, na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Relative Root Mean Squared Error.
}
\details{
Calculates the iqRMSE for a Predicted-Observed dataset by
normalizing the RMSE by the length of the inter-quartile range of
observations (percentiles 25th to 75th).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
iqRMSE(obs = X, pred = Y)
}
}
