% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land_cover}
\alias{land_cover}
\title{Binary Land Cover Data}
\format{
This data frame has 284 rows and the following 2 columns:
\describe{
\item{predicted}{predicted values}
\item{actual}{observed values}
}
}
\source{
\url{https://github.com/adriancorrendo/metrica}
}
\usage{
land_cover
}
\description{
This example dataset was obtained in 2022 over a small region in Kansas,
using visual interpretation.
The predicted values were the result of a Random Forest classification,
with a 70/30 \% split.
Values equal to 1 are associated to vegetation, and values equal to 0 are other type of land cover.
}
\keyword{datasets}
