% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCS.R
\name{LCS}
\alias{LCS}
\title{LCS}
\usage{
LCS(data = NULL, obs, pred, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Lack of correlation (LCS)component of the Mean Square Error (MSE).
}
\details{
Calculates the LCS for a Predicted-Observed dataset
following the MSE segregation suggested by Kobayashi and Salam
(2000).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
LCS(obs = X, pred = Y)
}
}
