% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize_phenology}
\alias{maize_phenology}
\title{Multi Class Phenology}
\format{
This data frame has 103 rows and the following 2 columns:
\describe{
\item{predicted}{predicted values}
\item{actual}{observed values}
}
}
\source{
\url{https://github.com/adriancorrendo/metrica}
}
\usage{
maize_phenology
}
\description{
This example data set includes maize phenology classes collected in Kansas during the 2018 growing season.
The predicted values were obtained using a Random Forest classifier with a 70/30 split.
The dataset includes 16 phenology stages from VT to R6.
For more information please visit
\doi{https://doi.org/10.3390/rs14030469}.
}
\keyword{datasets}
