% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPE.R
\name{MAPE}
\alias{MAPE}
\title{MAPE}
\usage{
MAPE(data = NULL, obs, pred, na.rm = TRUE)
}
\arguments{
\item{data}{(optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Mean Absolute Percentage Error.
}
\details{
Calculates the MAPE for a Predicted-Observed dataset.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
MAPE(obs = X, pred = Y)
}
}
