% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBE.R
\name{MBE}
\alias{MBE}
\title{MBE}
\usage{
MBE(obs, pred, na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Mean Bias Error.
}
\details{
Calculates the MBE of Predicted compared to Observed
values. Negative values indicate overestimation. Positive
values indicate underestimation. Moriasi et al. (2007).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
MBE(obs = X, pred = Y)
}
}
