% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcc.R
\name{mcc}
\alias{mcc}
\title{Matthews Correlation Coefficient}
\usage{
mcc(data = NULL, obs, pred, tidy = FALSE, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (character | factor).}

\item{pred}{Vector with predicted values (character | factor).}

\item{tidy}{Logical operator (TRUE/FALSE) to decide the type of return. TRUE
returns a data.frame, FALSE returns a list; Default : FALSE.}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
an object of class \code{numeric}.
}
\description{
It estimates the mcc for a nominal/categorical predicted-observed dataset.
}
\details{
The mcc it is also known as the phi-coefficient. It has gained
popularity within the machine learning community to summarize into a single
value the confusion matrix of a binary classification.

It is particularly useful when the number of observations belonging to each class
is uneven or imbalanced. It is characterized for being symmetric (i.e. no class
has more relevance than the other). It is bounded between -1 and 1.
The closer to 1 the better the classification performance.

The mcc is currently available only for the evaluation of binary cases (two classes).
For multiclass cases, \code{mcc} will produce a \code{NA} and display a warning.

For the formula and more details, see
\href{https://adriancorrendo.github.io/metrica/articles/available_metrics.html}{online-documentation}
}
\examples{
\donttest{
set.seed(123)
# Two-class
binomial_case <- data.frame(labels = sample(c("True","False"), 100, replace = TRUE), 
predictions = sample(c("True","False"), 100, replace = TRUE))

# Get mcc estimate for two-class case
mcc(data = binomial_case, obs = labels, pred = predictions, tidy = TRUE)
}
}
\references{
Chicco, D., Jurman, G. (2020)
The advantages of the Matthews correlation coefficient (MCC) over F1 score and
accuracy in binary classification evaluation.
\emph{BMC Genomics 21, 6 (2020).} \doi{10.1186/s12864-019-6413-7}
}
