% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_summary.R
\name{metrics_summary}
\alias{metrics_summary}
\title{metrics_summary}
\usage{
metrics_summary(
  data = NULL,
  obs,
  pred,
  orientation = "PO",
  na.rm = TRUE,
  metrics_list = NULL
)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{vector with observed values (numeric).}

\item{pred}{vector with predicted values (numeric).}

\item{orientation}{argument of class string specifying the axis
orientation to estimate slope(B1) and intercept(B0).
"PO" is for predicted vs observed, and "OP" for observed vs predicted.
Default is orientation = "PO".}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}

\item{metrics_list}{vector or list of specific selected metrics. Default is = NULL,
which will estimate all metrics available.}
}
\value{
an object of class \code{tibble}.
}
\description{
Scatter plot of Predictions and Observations.
}
\details{
Creates a table with one colum per metric.
}
\examples{
\dontrun{
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 10)
df <- data.frame(obs = X, pred = Y)
metrics_summary(df, obs = X, pred = Y)
}
}
