% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSE.R
\name{NSE}
\alias{NSE}
\title{NSE}
\usage{
NSE(obs, pred, na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Nash-Sutcliffe's Model Efficiency.
}
\details{
Calculates the NSE for a Predicted-Observed dataset
following Nash and Sutcliffe (1970).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 9)
NSE(obs = X, pred = Y)
}
}
