% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPB.R
\name{PPB}
\alias{PPB}
\title{PPB}
\usage{
PPB(obs, pred)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}
}
\value{
Element of class \code{numeric}.
}
\description{
Percentage Proportional Bias (PPB) component of the
Mean Square Error (MSE).
}
\details{
Calculates the PPB (\%) for a Predicted-Observed dataset
following the symmetric MSE decomposition suggested by
Correndo et al (2021).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
PPB(obs = X, pred = Y)
}
}
