% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRMSE.R
\name{RRMSE}
\alias{RRMSE}
\title{RRMSE}
\usage{
RRMSE(obs, pred, na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Relative Root Mean Square Error.
}
\details{
Calculates the RRMSE for a Predicted-Observed dataset.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
RRMSE(obs = X, pred = Y)
}
}
