% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSS.R
\name{RSS}
\alias{RSS}
\title{RSS}
\usage{
RSS(obs, pred, na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Residual Sum of Squares (R2).
}
\details{
Calculates the RSS for a Predicted-Observed dataset.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
RSS(obs = X, pred = Y)
}
}
