% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SB.R
\name{SB}
\alias{SB}
\title{SB}
\usage{
SB(data = NULL, obs, pred, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Squared bias (SB)component of the Mean Square
Error (MSE).
}
\details{
Calculates the SB for a Predicted-Observed dataset
following the MSE segregation suggested by Kobayashi and Salam
(2000).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 9)
SB(obs = X, pred = Y)
}
}
