% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.metrics.R
\name{scatter.metrics}
\alias{scatter.metrics}
\title{scatter.metrics}
\usage{
scatter.metrics(
  data = NULL,
  obs,
  pred,
  metrics = c("R2", "RMSE", "MAE", "NSE", "CCC")
)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{metrics}{Vector with up to 5 metrics to be displayed on the plot. Defaults to "R2","RMSE","MAE","NSE", and "CCC".}
}
\value{
Object of class \code{ggplot}
}
\description{
Creates a scatterplot with an embedded table displaying up to 5 metrics and the regression line
}
