% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.plot.R
\name{scatter.plot}
\alias{scatter.plot}
\title{scatter.plot}
\usage{
scatter.plot(obs, pred)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}
}
\value{
Object of class \code{ggplot}.
}
\description{
Scatter plot of Predictions and Observations.
}
\details{
Creates a scatter plot of predicted vs. observed values.
}
\examples{
\dontrun{
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 10)
scatter.plot(obs = X, pre = Y)
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{aes}}
}
