% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdev.R
\name{sdev}
\alias{sdev}
\title{sdev}
\usage{
sdev(x)
}
\arguments{
\item{x}{Vector with numeric elements.}
}
\value{
Element of class \code{numeric}.
}
\description{
Uncorrected Standard Deviation (uSD).
}
\details{
Calculates uSD from a vector x.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
sdev(x = X)
}
}
