% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDSD.R
\name{SDSD}
\alias{SDSD}
\title{SDSD}
\usage{
SDSD(obs, pred)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}
}
\value{
Element of class \code{numeric}.
}
\description{
Square difference between standard deviations (SDSD)
component of the Mean Square Error (MSE).
}
\details{
Calculates the SDSD for a Predicted-Observed dataset
following the MSE segregattion suggested by Kobayashi and Salam
(2000).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
SDSD(obs = X, pred = Y)
}
}
