% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sorghum}
\alias{sorghum}
\title{Sorghum grain number}
\format{
This data frame has 36 rows and the following 2 columns:
\describe{
\item{pred}{predicted values}
\item{obs}{observated values}
}
}
\source{
\url{https://github.com/adriancorrendo/metrica}
}
\usage{
sorghum
}
\description{
This example dataset is a set of APSIM simulations of
sorghum grain number (x1000 grains per squared meter), which exhibits
both low accuracy and low precision as it represents a model still
under development. The experimental trials come from 6 site-years in
1 country (Australia).
These data correspond to the latest, up-to-date, documentation and
validation of version number 2020.03.27.4956. Data available at:
\url{https://doi.org/10.7910/DVN/EJS4M0}.
Further details can be found at the official APSIM Next Generation
website:
\url{https://APSIMnextgeneration.netlify.app/modeldocumentation/}
}
\keyword{datasets}
