% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec}
\alias{spec}
\title{Specificity}
\usage{
spec(data = NULL, obs, pred, tidy = FALSE, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (character | factor).}

\item{pred}{Vector with predicted values (character | factor).}

\item{tidy}{Logical operator (TRUE/FALSE) to decide the type of return. TRUE
returns a data.frame, FALSE returns a list; Default : FALSE.}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
an object of class \code{numeric}.
}
\description{
It estimates the specificity for a nominal/categorical predicted-observed dataset.
}
\details{
The specificity (or sensitivity) is a non-normalized coefficient that represents the
ratio between the correctly predicted cases to the total number of actual observations
that belong to a given class.

For binomial cases, specificity  =  True positives/total number of actual positives = TP/(TP + FN)

It is bounded between 0 and 1.
The closer to 1 the better. Values towards zero indicate low performance.
It can be estimated for each particular class or at a global level.
For the formula and more details, see
\href{https://adriancorrendo.github.io/metrica/articles/available_metrics.html}{online-documentation}
}
\examples{
\donttest{
set.seed(123)
# Two-class
binomial_case <- data.frame(labels = sample(c("True","False"), 100, 
replace = TRUE), predictions = sample(c("True","False"), 100, replace = TRUE))
# Multi-class
multinomial_case <- data.frame(labels = sample(c("Red","Blue", "Green"), 100, 
replace = TRUE), predictions = sample(c("Red","Blue", "Green"), 100, replace = TRUE)    )

# Get specificity estimate for two-class case
spec(data = binomial_case, obs = labels, pred = predictions, tidy = TRUE)

# Get specificity estimate for each class for the multi-class case
spec(data = multinomial_case, obs = labels, pred = predictions, tidy = TRUE)

# Get specificity estimate for the multi-class case at a global level
spec(data = multinomial_case, obs = labels, pred = predictions, tidy = TRUE)
}
}
\references{
Ting K.M. (2017)
Precision and Recall.
\emph{In: Sammut C., Webb G.I. (eds) Encyclopedia of Machine Learning and Data Mining.}
\emph{Springer, Boston, MA.} \doi{10.1007/978-1-4899-7687-1_659}
(2017) Sensitivity.
\emph{In: Sammut C., Webb G.I. (eds) Encyclopedia of Machine Learning and Data Mining.}
\emph{Springer, Boston, MA.} \doi{10.1007/978-1-4899-7687-1_751}
}
