% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSx.R
\name{SSx}
\alias{SSx}
\title{SSx}
\usage{
SSx(x)
}
\arguments{
\item{x}{Vector with numeric elements.}
}
\value{
Element of class \code{numeric}.
}
\description{
Sum of Squares (SS) or Uncorrected Variance (not n-1).
}
\details{
Calculates the SS from a vector x.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
SSx(x = X)
}
}
