% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.R
\name{TSS}
\alias{TSS}
\title{TSS}
\usage{
TSS(data = NULL, obs, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Total Sum of Squares (R2).
}
\details{
Calculates the TSS for a Predicted-Observed dataset.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
TSS(obs = X, pred = Y)
}
}
