% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uc.R
\name{Uc}
\alias{Uc}
\title{Uc}
\usage{
Uc(obs, pred, na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Lack of consistency (Uc). Proportion of the total sum
of squares related to the proportional bias.
}
\details{
Calculates the Uc (\%) for a Predicted-Observed dataset
following the sum of squares segregattion suggested by Smith and
Rose (1995) also known as Theil's partial inequalities.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
Uc(obs = X, pred = Y)
}
}
