% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.u.R
\name{var.u}
\alias{var.u}
\title{var.u}
\usage{
var.u(data = NULL, x, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{x}{Vector with numeric elements.}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Uncorrected Sample Variance (var.u, not n-1).
}
\details{
Calculates the var.u of a vector x.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
var.u(x = X)
}
}
