% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wheat}
\alias{wheat}
\title{Wheat grain nitrogen}
\format{
This data frame has 137 rows and the following 2 columns:
\describe{
\item{pre}{predicted values}
\item{obs}{observated values}
}
}
\source{
\url{https://github.com/adriancorrendo/simob}
}
\usage{
wheat
}
\description{
This example dataset is a set of APSIM simulations of
wheat grain N (grams per squared meter), which presents both high
accuracy and high precision. The experimental trials come from 11
site-years in 4 countries (Australia, Ethiopia, New Zealand, and
Turkey).
These data correspond to the latest, up-to-date, documentation and
validation of version number 2020.03.27.4956. Data available at:
\url{https://doi.org/10.7910/DVN/EJS4M0}.
Further details can be found at the official APSIM Next Generation
website:
\url{https://apsimnextgeneration.netlify.app/modeldocumentation/}
}
\keyword{datasets}
