% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xa.R
\name{Xa}
\alias{Xa}
\title{Xa}
\usage{
Xa(obs, pred, na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{double}.
}
\description{
Accuracy component of Concordance Correlation.
}
\details{
Calculates the Xa for a Predicted-Observed dataset.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
Xa(obs = X, pred = Y)
}
}
