% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AC.R
\name{AC}
\alias{AC}
\title{AC}
\usage{
AC(obs, pred, na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}
}
\value{
Element of class \code{numeric}.
}
\description{
Ji and Gallo's Agreement Coefficient (AC).
}
\details{
Calculates the AC for a Predicted-Observed dataset
following Ji and Gallo (2006).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 9)
AC(obs = X, pred = Y)
}
}
