% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland_altman_plot.R
\name{bland_altman_plot}
\alias{bland_altman_plot}
\title{bland_altman_plot}
\usage{
bland_altman_plot(data = NULL, obs, pred, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values}
}
\value{
and object of class \code{ggplot}.
}
\description{
Bland-Altman plot.
}
\details{
Creates a scatter plot of the difference between
observed and predicted values (obs-pred) vs. observed values.
}
\examples{
\dontrun{
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 10)
bland_altman_plot(obs = X, pred = Y)
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{aes}}
}
