% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d1.R
\name{d1}
\alias{d1}
\title{d1}
\usage{
d1(obs, pred, na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Modified Willmott's Index of Agreement (d1).
}
\details{
Calculates the d1 for a Predicted-Observed dataset
following Willmott et al. (2012).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 9)
KGE(obs = X, pred = Y)
}
}
