% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d1r.R
\name{d1r}
\alias{d1r}
\title{d1r}
\usage{
d1r(obs, pred, na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Refined Willmott's Index of Agreement (d1r).
}
\details{
Calculates the d1e for a Predicted-Observed dataset
following Willmott et al. (2012), which modifies the denominator
(potential error) when normalizing the mean absolute error
(numerator).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 9)
d1r(obs = X, pred = Y)
}
}
