% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.apsim.db.R
\name{import.apsim.db}
\alias{import.apsim.db}
\title{import.apsim.db}
\usage{
import.apsim.db(filename = "", folder = ".", value = "report", simplify = TRUE)
}
\arguments{
\item{filename}{file name including the file extension ("*.db"), as a string ("").}

\item{folder}{source folder/directory containing the file, as a string ("").}

\item{value}{either \sQuote{report}, \sQuote{all} (list) or user-defined for a specific report.}

\item{simplify}{if TRUE will attempt to simplify multiple reports into a single data.frame.
If FALSE it will return a list.}
}
\value{
An object of class \code{data.frame}, but it depends on the argument \sQuote{value} above
}
\description{
Imports data from SQLite databases (*.db) generated by APSIM NextGen
}
\note{
Adapted from apsimx to maintain compatibility. For reference, we also recommend
to check and use the function apsimx::read.apsimx() as an alternative.
(source: https://github.com/femiguez/apsimx by F. Miguez).

NOTE: if there is one single report it will return a data.frame.
If there are multiple reports, it will attempt to merge them into a data frame.
If not possible it will return a list with names corresponding to the
table report names. It is also possible to select a specific report from several
available by selecting \sQuote{value = ReportName}, where \sQuote{ReportName} is the name
of the specific report that should be returned.
If you select \sQuote{all} it will return all the components in the data base also as a list.
}
\seealso{
\code{\link[apsimx]{read.apsimx}},\code{\link[RSQLite]{SQLite}},\code{\link[RSQLite]{dbListTables}},
\code{\link[DBI]{dbGetQuery}}
}
