% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.apsim.out.R
\name{import.apsim.out}
\alias{import.apsim.out}
\title{import.apsim.out}
\usage{
import.apsim.out(filepath)
}
\arguments{
\item{filepath}{Argument to specify the location path to the *.out file.}
}
\value{
Element of class \code{data.frame}.
}
\description{
Function to import *.out files generated by APSIM Classic.
}
\details{
This function will generate a data frame object from an APSIM Classic *.out file.
}
\note{
it will create a data frame object, one file at a time.
}
\examples{
\dontrun{
# Pull the git repo and use the following filepath
path <- "tests/testthat/examples/soybean.out"
soybean.out <- import.apsim.out(filepath = path)
head(soybean.out)
}
}
\seealso{
\code{\link[apsimx]{read.apsim}}
}
