% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KGE.R
\name{KGE}
\alias{KGE}
\title{KGE}
\usage{
KGE(obs, pred, na_rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}
}
\value{
Element of class \code{numeric}.
}
\description{
Kling-Gupta Model Efficiency (KGE).
}
\details{
Calculates the KGE for a Predicted-Observed dataset
following Kling et al. (2012).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 9)
KGE(obs = X, pred = Y)
}
}
