% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KGE.R
\name{KGE}
\alias{KGE}
\title{KGE}
\usage{
KGE(obs, pred, na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Kling-Gupta Model Efficiency (KGE).
}
\details{
Calculates the KGE for a Predicted-Observed dataset
following Kling et al. (2012).
Kling, H., M. Fuchs, and M. Paulin (2012), Runoff conditions in the
upper Danube basin under an ensemble of climate change scenarios,
Journal of Hydrology 424-425, 264-277,
doi:10.1016/j.jhydrol.2012.01.011.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 9)
KGE(obs = X, pred = Y)
}
}
