% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBE.R
\name{MBE}
\alias{MBE}
\title{MBE}
\usage{
MBE(obs, pre)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pre}{Vector with predicted values (numeric).}
}
\value{
Element of class \code{double}.
}
\description{
Mean Bias Error.
}
\details{
Calculates the MBErr of Predicted compared to Observed
values.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 9)
MBE(obs = X, pre = Y)
}
}
