% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrica-package.R
\docType{package}
\name{metrica-package}
\alias{metrica}
\alias{metrica-package}
\title{metrica: Error Metrics for Predicted-Observed Datasets}
\description{
Metrica offers users of simulation models a complete toolbox with error metrics accounting for different aspects of the agreement between predicted and observed values.
}
\author{
\strong{Maintainer}: Adrian Correndo \email{correndo@ksu.edu} (\href{https://orcid.org/0000-0002-4172-289X}{ORCID})

Authors:
\itemize{
  \item Luiz Moro Rosso \email{lhmrosso@ksu.edu} (\href{https://orcid.org/0000-0002-8642-911X}{ORCID})
  \item Dean Holzworth \email{dean.holzworth@csiro.au}
  \item Ignacio Ciampitti \email{ciampitti@ksu.edu} (\href{https://orcid.org/0000-0001-9619-5129}{ORCID})
}

}
\keyword{internal}
