% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrica-package.R
\docType{package}
\name{metrica-package}
\alias{metrica}
\alias{metrica-package}
\title{metrica: Prediction Performance Metrics}
\description{
A compilation of more than 40 functions designed to evaluate prediction performance of point-forecast models accounting for different aspects of the agreement between predicted and observed values using error metrics (e.g. MAE, RMSE), model efficiencies (e.g. NSE, KGE), indices of agreement (e.g. d, RAC), goodness of fit (e.g. r, R2), concordance correlation (e.g. CCC), error decomposition (e.g. lack of accuracy-precision), and plots the visualize agreement. For more details visit the vignettes <https://adriancorrendo.github.io/metrica/>
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adriancorrendo.github.io/metrica/}
  \item Report bugs at \url{https://github.com/adriancorrendo/metrica/issues}
}

}
\author{
\strong{Maintainer}: Adrian A. Correndo \email{correndo@ksu.edu} (\href{https://orcid.org/0000-0002-4172-289X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adrian A. Correndo \email{correndo@ksu.edu} (\href{https://orcid.org/0000-0002-4172-289X}{ORCID})
  \item Luiz H. Moro Rosso \email{lhmrosso@ksu.edu} (\href{https://orcid.org/0000-0002-8642-911X}{ORCID})
  \item Rai Schwalbert \email{rai.schwalbert@hotmail.com} (\href{https://orcid.org/0000-0001-8488-7507}{ORCID})
  \item Carlos Hernandez \email{carlosh92@ksu.edu} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})
  \item Leonardo M. Bastos \email{leonardombastos@gmail.com} (\href{https://orcid.org/0000-0001-8958-6527}{ORCID})
  \item Dean Holzworth \email{dean.holzworth@csiro.au}
  \item Ignacio A. Ciampitti \email{ciampitti@ksu.edu} (\href{https://orcid.org/0000-0001-9619-5129}{ORCID})
}

}
\keyword{internal}
