% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrica_plot.R
\name{metrica_plot}
\alias{metrica_plot}
\title{metrica_plot}
\usage{
metrica_plot(obs, pre)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pre}{Vector with predicted values (numeric).}
}
\value{
Object of class \code{ggplot}.
}
\description{
Scatter plot of Predictions and Observations.
}
\details{
Creates a scatter plot of sim vs. obs.
}
\examples{
\dontrun{
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 10)
metrica_plot(obs = X, pre = Y)
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{aes}}
}
