% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.summary.R
\name{metrics.summary}
\alias{metrics.summary}
\title{metrics.summary}
\usage{
metrics.summary(obs, pred, na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Object of class \code{tibble}.
}
\description{
Scatter plot of Predictions and Observations.
}
\details{
Creates a table with one colum per metric.
}
\examples{
\dontrun{
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 10)
metrics.summary(obs = X, pre = Y)
}
}
