% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLA.R
\name{MLA}
\alias{MLA}
\title{MLA}
\usage{
MLA(data = NULL, obs, pred, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Mean Lack of Accuracy (MLA) component of the Mean
Square Error (MSE).
}
\details{
Calculates the MLA for a Predicted-Observed dataset
following the symmetric MSE decomposition suggested by
Correndo et al (2021).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
df <- data.frame(obs = X, pred = Y)
MLA(df, obs = X, pred = Y)
}
}
