% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLP.R
\name{MLP}
\alias{MLP}
\title{MLP}
\usage{
MLP(obs, pred, na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Mean Lack of Precision (MLP) component of the Mean
Square Error (MSE).
}
\details{
Calculates the MLP for a Predicted-Observed dataset
following the MSE decomposition suggested by Correndo et al (2021).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
MLP(obs = X, pred = Y)
}
}
