% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.R
\name{MSE}
\alias{MSE}
\title{MSE}
\usage{
MSE(data, obs, pre)
}
\arguments{
\item{data}{Dataframe with observed and predicted vectors.}

\item{obs}{Vector with observed values (numeric).}

\item{pre}{Vector with predicted values (numeric).}
}
\value{
Element of class \code{double}.
}
\description{
Mean Square Error.
}
\details{
Calculates the MSE comparing Predicted and Observed
values.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 9)
MSE(obs = X, pre = Y)
}
}
