% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBE.R
\name{PBE}
\alias{PBE}
\title{PBE}
\usage{
PBE(obs, pred)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}
}
\value{
Element of class \code{numeric}.
}
\description{
Percentage Bias.
}
\details{
Calculates the PBE (\%) for a Predicted-Observed dataset.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
PBE(obs = X, pred = Y)
}
}
