% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAC.R
\name{RAC}
\alias{RAC}
\title{RAC}
\usage{
RAC(obs, pred)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}
}
\value{
Element of class \code{numeric}.
}
\description{
Robinson's Agreement Coefficient.
}
\details{
Calculates the RAC for a Predicted-Observed dataset
following Robinson (1957,1959).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
RAC(obs = X, pred = Y)
}
}
