% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAE.R
\name{RAE}
\alias{RAE}
\title{RAE}
\usage{
RAE(data = NULL, obs, pred, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Relative Abosulte Error.
}
\details{
Calculates the RAE for a Predicted-Observed dataset as the
the ratio between the sum of absolute residuals (error of
predictions with respect to observations) and the sum of absolute
errors of observations with respect to its mean.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
RAE(obs = X, pred = Y)
}
}
