% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{RMSE}
\usage{
RMSE(obs, pred)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}
}
\value{
Element of class \code{numeric}.
}
\description{
Root Mean Square Error.
}
\details{
Calculates the RMSE for a Predicted-Observed dataset.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
RMSE(obs = X, pred = Y)
}
}
