% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{RMSE}
\usage{
RMSE(obs, pre)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pre}{Vector with predicted values (numeric).}
}
\value{
Element of class \code{double}.
}
\description{
Root Mean Square Error.
}
\details{
Calculates the RMSE comparing Predicted and Observed
values.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 9)
RMSE(obs = X, pre = Y)
}
}
