% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_RRMSE.R
\name{RRMSE}
\alias{RRMSE}
\title{Relative Root Mean Squared Error (RMSE)}
\usage{
RRMSE(data = NULL, obs, pred, tidy = FALSE, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{tidy}{Logical operator (TRUE/FALSE) to decide the type of return. TRUE
returns a data.frame, FALSE returns a list; Default : FALSE.}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
an object of class \code{numeric} within a \code{list} (if tidy = FALSE) or within a
\verb{data frame} (if tidy = TRUE).
}
\description{
It estimates the RRMSE for a continuous predicted-observed dataset.
}
\details{
The RRMSE normalizes the Root Mean Squared Error (RMSE) by the mean
of observations. It goes from 0 to infinity. The lower the better the prediction performance.
In literature, it can be also found as NRMSE (normalized root mean squared error).
However, here we use RRMSE since several other alternatives to
"normalize" the RMSE exist (e.g., RSR, iqRMSE).
For the formula and more details, see \href{https://adriancorrendo.github.io/metrica/articles/available_metrics_regression.html}{online-documentation}
}
\examples{
\donttest{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
RRMSE(obs = X, pred = Y)
}
}
