% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSE.R
\name{RSE}
\alias{RSE}
\title{RSE}
\usage{
RSE(obs, pred, na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Relative Squared Error.
}
\details{
Calculates the RSE for a Predicted-Observed dataset. It is
the ratio between the residual sum of squares (RSS, error of
predictions with respect to observations) and the total sum of
squares (TSS, error of observations with respect to its mean).
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
RSE(obs = X, pred = Y)
}
}
