% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.plot.R
\name{scatter.plot}
\alias{scatter.plot}
\title{scatter.plot}
\usage{
scatter.plot(obs, pred, orientation = "PO", na.rm = TRUE)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{orientation}{Argument of class string specifying the axis
orientation, PO for predicted vs observed, and OP for
observed vs predicted. Default is orientation = "PO".}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Object of class \code{ggplot}.
}
\description{
Scatter plot of Predictions and Observations.
}
\details{
Creates a scatter plot of predicted vs. observed values.
}
\examples{
\dontrun{
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 10)
scatter.plot(obs = X, pre = Y)
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{aes}}
}
