% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMAPE.R
\name{SMAPE}
\alias{SMAPE}
\title{SMAPE}
\usage{
SMAPE(obs, pred)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}
}
\value{
Element of class \code{numeric}.
}
\description{
Symmetric Mean Absolute Percentage Error.
}
\details{
Calculates the SMAPE for a Predicted-Observed dataset.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
MAPE(obs = X, pred = Y)
}
}
