% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.plot.R
\name{tiles.plot}
\alias{tiles.plot}
\title{tiles.plot}
\usage{
tiles.plot(obs, pred, bins = 10)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{bins}{Argument of class numeric specifying the number
of tile color groups.}
}
\value{
Object of class \code{ggplot}.
}
\description{
Scatter plot of Predictions and Observations.
}
\details{
Creates a scatter plot of sim vs. obs.
}
\examples{
\dontrun{
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 10)
tiles.plot(obs = X, pre = Y)
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{aes}}
}
