% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.plot.R
\name{tiles.plot}
\alias{tiles.plot}
\title{tiles.plot}
\usage{
tiles.plot(
  data = NULL,
  obs,
  pred,
  bins = 10,
  orientation = "PO",
  print.metrics = FALSE,
  metrics.list = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{bins}{Argument of class numeric specifying the number}

\item{orientation}{Argument of class string specifying the axis
orientation, PO for predicted vs observed, and OP for
observed vs predicted. Default is orientation = "PO".}

\item{print.metrics}{boolean TRUE/FALSE to embed metrics in the plot. Default is FALSE.}

\item{metrics.list}{vector or list of selected metrics to print on the plot.}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Object of class \code{ggplot}.
}
\description{
Scatter plot of Predictions and Observations.
}
\details{
Creates a scatter plot of sim vs. obs.
}
\examples{
\dontrun{
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- rnorm(n = 100, mean = 0, sd = 10)
tiles.plot(obs = X, pre = Y)
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{aes}}
}
