% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ub.R
\name{Ub}
\alias{Ub}
\title{Ub}
\usage{
Ub(data = NULL, obs, pred, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Proportion of the total sum of squares related to the
mean bias (Ub).
}
\details{
Calculates the Ub for a Predicted-Observed dataset
following the sum of squares segregation suggested by Smith and
Rose (1995) also known as Theil's partial inequalities.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
Ub(obs = X, pred = Y)
}
}
