% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ue.R
\name{Ue}
\alias{Ue}
\title{Ue}
\usage{
Ue(obs, pred)
}
\arguments{
\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}
}
\value{
Element of class \code{numeric}.
}
\description{
Lack of consistency (Ue). Proportion of the total sum
of squares related to the proportional bias.
}
\details{
Calculates the Ue (\%) for a Predicted-Observed dataset
following the sum of squares segregattion suggested by Smith and
Rose (1995) also known as Theil's partial inequalities.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
Ue(obs = X, pred = Y)
}
}
