% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSD.R
\name{uSD}
\alias{uSD}
\title{uSD}
\usage{
uSD(data = NULL, x, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{x}{Vector with numeric observed or predicted values.}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
Element of class \code{numeric}.
}
\description{
Uncorrected Standard Deviation (uSD) of observed or predicted values.
}
\details{
Calculates uSD from a vector x.
}
\examples{
\dontrun{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
uSD(x = X)
}
}
