% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xa.R
\name{Xa}
\alias{Xa}
\title{Accuracy Component (Xa) of CCC}
\usage{
Xa(data = NULL, obs, pred, na.rm = TRUE)
}
\arguments{
\item{data}{(Optional) argument to call an existing data frame containing the data.}

\item{obs}{Vector with observed values (numeric).}

\item{pred}{Vector with predicted values (numeric).}

\item{na.rm}{Logic argument to remove rows with missing values
(NA). Default is na.rm = TRUE.}
}
\value{
an object of class \code{double}.
}
\description{
It estimates the Xa component for the calculation of the Concordance
Correlation Coefficient (CCC) following Lin (1989).
}
\details{
The Xa measures accuracy of prediction. It goes from 0 (completely inaccurate) to 1 (perfectly accurate).
It is used to adjust the precision measured by the correlation coefficient (r)
in order to evaluate agreement through the CCC.
For the formula and more details, see \href{https://adriancorrendo.github.io/metrica/articles/available_metrics.html}{online-documentation}
}
\examples{
\donttest{
set.seed(1)
X <- rnorm(n = 100, mean = 0, sd = 10)
Y <- X + rnorm(n=100, mean = 0, sd = 3)
Xa(obs = X, pred = Y)
}
}
\references{
Lin (1989).
A concordance correlation coefficient to evaluate reproducibility.
\emph{Biometrics 45 (1), 255–268.} \doi{10.2307/2532051}
}
